[virt] Move interval_timer.{h,cpp} to virt/.
