+++ src/virt/interval_timer.h
